fix: layer: clean up layer info
emplacing a value into an unordered map does not override, so if the swapchain handles were identical (which they are commonly), then it would try to continue with the swapchain images from the old swapchain.

ironically, the swapchain images are also mostly identical. in fact, when using validation layers they are intentionally-unintentionally identical.. so I never caught this bug.

either way, this fixes a bunch of weird behavior:
- swapchain creation failing
- the application hanging on a present
- the infamous "assertion failed!" wine warning
- segmentation faults in the NVIDIA driver